refactored some redux code, added index build to configureStore and redux
