+++ src/data/actions/search-actions.js
@@ -1,0 +1,6 @@
+const setIndex = index => ({
+  index,
+  type: 'SET_INDEX',
+});
+
+export default setIndex;
