+++ src/data/configureStore.js
@@ -1,8 +1,10 @@
 import { createStore } from 'redux';
 
 import { setImages } from './actions/image-actions';
+import setIndex from './actions/search-actions';
 import imageStore from './store';
 import { getImages } from '../image/support/retriever';
+import buildIndex from '../search';
 
 const configureStore = () => {
   const store = createStore(imageStore);
@@ -11,7 +13,13 @@ const configureStore = () => {
     store.dispatch(setImages(images));
   });
 
+  buildIndex('Images', (index) => {
+    store.dispatch(setIndex(index));
+  });
+
   return store;
 };
+
+configureStore();
 
 export default configureStore;
