+++ src/data/store.js
@@ -1,64 +1,12 @@
 import { combineReducers } from 'redux';
 
+import { image, images, previewImageUrl } from './reducers/image-store';
+import { searchIndex, searchOpen, searchQuery } from './reducers/search-store';
+
 const canUpload = (state = false, action) => {
   switch (action.type) {
     case 'SET_UPLOAD':
       return action.upload;
-
-    default:
-      return state;
-  }
-};
-
-const image = (state = {
-  name: '',
-  width: 0,
-  height: 0,
-}, action) => {
-  switch (action.type) {
-    case 'SET_IMAGE_HEIGHT': {
-      return {
-        ...state,
-        height: action.height,
-      };
-    }
-
-    case 'SET_IMAGE_WIDTH': {
-      return {
-        ...state,
-        width: action.width,
-      };
-    }
-
-    case 'SET_IMAGE_FILE_NAME': {
-      return {
-        ...state,
-        name: action.name,
-      };
-    }
-
-    default:
-      return state;
-  }
-};
-
-const images = (state = [], action) => {
-  switch (action.type) {
-    case 'ADD_IMAGE':
-      return [action.image, ...state];
-
-    case 'SET_IMAGES':
-      return action.images;
-
-    default:
-      return state;
-  }
-};
-
-const previewImageUrl = (state = '', action) => {
-  switch (action.type) {
-    case 'SET_PREVIEW_IMAGE_URL':
-      return action.url;
 
     default:
       return state;
@@ -98,35 +46,13 @@ const previewLabels = (state = [], action) => {
   }
 };
 
-const searchOpen = (state = false, action) => {
-  switch (action.type) {
-    case 'SET_SEARCH_OPEN':
-      return action.open;
-
-    default:
-      return state;
-  }
-};
-
-const searchQuery = (state = '', action) => {
-  switch (action.type) {
-    case 'CLEAR_SEARCH_QUERY':
-      return '';
-
-    case 'SET_SEARCH_QUERY':
-      return action.query;
-
-    default:
-      return state;
-  }
-};
-
 export default combineReducers({
   canUpload,
   image,
   images,
   previewImageUrl,
   previewLabels,
+  searchIndex,
   searchOpen,
   searchQuery,
 });
